.azuredevops/templates/requirements/requirements_ci.txt
azure-identity==1.*
azure-keyvault-secrets==4.*
azure-storage-file-datalake==12.*
black
databricks-connect==15.*
databricks-sql-connector==2.*
dbx
furo
myst-parser
nest_asyncio
pandas
pip
pyarrow
pylint==3.3.*
pytest-asyncio
pytest-cov
pytest-nunit
pytest-xdist
pytest==7.*
python-dotenv
rich
setuptools
sphinx-click
Sphinx>2.0.0S
strictyaml
typer
wheel


Click==8.1.* 
aiohttp
requests